refactor search component to compositions
