+++ frontend/compositions/search/search-panel.tsx
